+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-
+* Fixed an issue in **update-release-notes** where suggestion fix failed in validation.
 
 # 1.6.1
 * Added the '--use-packs-known-words' argument to the **doc-review** command