+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
+
 * Fixed an issue in **update-release-notes** where suggestion fix failed in validation.
+* Fixed a bug where `.env` files didn't load. @nicolas-rdgs
 
 # 1.6.1
 * Added the '--use-packs-known-words' argument to the **doc-review** command