+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Fixed an issue in **update-release-notes** where suggestion fix failed in validation.
 * Fixed a bug where `.env` files didn't load. @nicolas-rdgs
-* Fixed an issue in **update-release-notes** where certain words failed in validation.
+* Fixed an issue in **review-release-notes** where certain words failed in validation.
 
 # 1.6.1
 * Added the '--use-packs-known-words' argument to the **doc-review** command